requirements.txt
arabic-reshaper 
python-bidi 
pandas 
openpyxl 
emoji 
joblib 
